Added a pytest util file for sharing common functions
