+++ realsense2_camera/test/pytest_rs_utils.py
@@ -1,0 +1,80 @@
+# Copyright 2023 Intel Corporation. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import os
+import sys
+
+assert os.getenv("COLCON_PREFIX_PATH")!=None,"COLCON_PREFIX_PATH was not set" 
+sys.path.append(os.getenv("COLCON_PREFIX_PATH")+'/realsense2_camera/share/realsense2_camera/launch')
+import rs_launch
+
+'''
+get the default parameters from the launch script so that the test doesn't have to
+get updated for each change to the parameter or default values 
+'''
+def get_default_params():
+    params = {}
+    for param in rs_launch.configurable_parameters:
+        params[param['name']] = param['default']
+    return params
+
+'''
+function taken from rs_launch to kill the camera node. kept as a local copy so that when the template is 
+used, it can be changed to kill say, a particular node alone depending on the test scenario
+'''
+
+def kill_realsense2_camera_node():
+    cmd = "kill -s INT $(ps aux | grep '[r]ealsense2_camera_node' | awk '{print $2}')"
+    os.system(cmd)
+
+'''
+get the default parameters from the launch script so that the test doesn't have to
+get updated for each change to the parameter or default values 
+'''
+
+def get_default_params():
+    params = {}
+    for param in rs_launch.configurable_parameters:
+        params[param['name']] = param['default']
+    return params
+
+''' 
+The format used by rs_launch.py and the LuanchConfiguration yaml files are different,
+so the params reused from the rs_launch has to be reformated to be added to yaml file.
+'''
+def convert_params(params):
+    cparams = {}
+    def strtobool (val):
+        val = val.lower()
+        if val == 'true':
+            return True
+        elif val == 'false':
+            return False
+        else:
+            raise ValueError("invalid truth value %r" % (val,))
+    for key, value in params.items():
+        try:
+            cparams[key] = int(value)
+        except ValueError:
+            try:
+                cparams[key] = float(value)
+            except ValueError:
+                try:
+                    cparams[key] = strtobool(value)
+                except ValueError:
+                    cparams[key] = value.replace("'","")
+    return cparams
+
+def get_params_string_for_launch(params):
+    params_str = ' '.join(["" if params[key]=="''" else key + ':=' + params[key] for key in sorted(params.keys())])
+    return params_str
